+++ src/sparseml/utils/fsdp/helpers.py
@@ -29,12 +29,9 @@ from sparseml.utils.pytorch import set_layer
 
 __all__ = [
     "is_fsdp_model",
-    "fix_fsdp_module_name",
     "unwrap_and_export_model",
     "save_pretrained_fsdp",
 ]
-
-FSDP_WRAPPER_NAME = "_fsdp_wrapped_module."
 
 
 def is_fsdp_model(model: Module) -> bool:
@@ -48,16 +45,6 @@ def is_fsdp_model(model: Module) -> bool:
         return False
 
     return isinstance(model, FullyShardedDataParallel)
-
-
-def fix_fsdp_module_name(name: str) -> str:
-    """
-    Remove FSDP wrapper prefixes from a module name
-
-    :param name: name to strip
-    :return: stripped name
-    """
-    return name.replace(FSDP_WRAPPER_NAME, "")
 
 
 def unwrap_and_export_model(model, accelerator, output_dir, tokenizer):
